feat(hazards): expand the whole panel into a wide analysis view
ALL-5673. "Expand" now opens the full panel as a wide dashboard instead of only
the two charts: KPIs across the top, filters + the basin ranking in the left rail
(the spatial "where"), and the depth + affected-area charts on the right (the
temporal "how"). Metrics that belong together finally sit together, with room
the ~300px sidebar can't give. Filters in the view write the same URL state, so
closing it leaves them applied to the map; drilling into a basin refines the
KPIs, header, and both charts.

Structure (no computation duplication, no import cycle): the old pop-out's chart
body is extracted to DisplacementDetailCharts; DisplacementAnalysisDialog is a
pure layout shell that takes already-rendered slots (filters/KPIs/ranking/charts)
so it imports nothing from the stats components that open it; DisplacementLayerCharts
builds the KPI + ranking nodes once and reuses them in both the sidebar column and
the analysis view, so the two can't drift. The standalone charts-only dialog is
removed (the analysis view supersedes it).

Deferred (noted): an in-view surface switch, and a Rate "Expand" (Rate has no
time-series, so its sidebar KPIs+ranking already stand alone) — both fast-follows.

Reviewed: built to the scientist-endorsed mockup. Verified: tsc clean, vitest
53/53, eslint clean (pre-existing warnings only), build green; live-checked — the
wide layout renders, drill-in updates KPIs+header+charts (Cedar City → 67.7 mi²,
13 in, 1 basin), and the sidebar is unchanged.
